backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.42.0
python-dotenv==1.0.1
pydantic==2.10.5
httpx==0.28.1
sec-api==1.0.17
pypdf==5.1.0
